requirements.txt
# Marketing Agent Pipeline - Dependencies

# AI Agent Framework
crewai[tools]>=0.100.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20

# Search & Research
tavily-python>=0.3.0
exa-py>=1.0.0
requests>=2.31.0

# Social Media APIs
python-twitter-v2>=0.9.0
instagrapi>=2.0.0

# Email Marketing
mailerlite-api-client>=1.0.0

# Scheduling & Workflow
schedule>=1.2.0
celery>=5.3.0
redis>=5.0.0

# Web Framework (for API + webhooks)
fastapi>=0.100.0
uvicorn>=0.25.0

# MCP Server (remote access)
mcp>=1.0.0
starlette>=0.36.0
sse-starlette>=1.8.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.0.0
